refactor(api): PR-6.5 commit 2 — train.zip / bundle.zip + 跨项目 import 6 routes
server.py 2072 → 1795 行（−277），projects 域第二刀（exports / imports）。

  api/routers/projects/exports.py    6 routes  train.zip / bundle.zip /
                                               export-bundle / import-bundle
                                               (path + upload) / import-train
                                               + _bundle_import_payload +
                                               _import_bundle_from_path helpers
  api/schemas/exports.py             2 BaseModel  BundleOptionsBody (含 to_options)
                                                  + BundleImportBody (exactly_one
                                                  validator)

行为零变更：临时文件 tempfile.NamedTemporaryFile + BackgroundTasks 清理 +
SSE publish version_train_zip_ready/_failed + version_bundle_zip_ready/_failed
全保留。

回归：PR-1 安全网 152 + project_endpoints 49 + ... = **203/203 全绿**。
route snapshot 不变（160 routes）。
